requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
pandas==2.1.3
numpy==1.26.4
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
json5==0.9.14
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
reportlab==4.4.3
google-generativeai==0.7.2
setuptools>=68.0.0
wheel>=0.41.0
